agents/herald/__init__.py
from pathlib import Path
from dotenv import load_dotenv
load_dotenv(Path(__file__).resolve().parent.parent.parent / '.env')
# agents/herald/__init__.py
# Capa HERALD — Publicación: OLYMPUS, RAPID, MERCURY

from agents.herald.olympus import OLYMPUS
from agents.herald.rapid import RAPID
from agents.herald.mercury import MERCURY

__all__ = ["OLYMPUS", "RAPID", "MERCURY"]

