requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
groq==0.13.0
httpx==0.28.0
feedparser==6.0.11
edge-tts==6.1.12
moviepy==1.0.3
Pillow==11.0.0
matplotlib==3.9.3
python-telegram-bot==21.9
python-dotenv==1.0.1
pyyaml==6.0.2
rich==13.9.4
requests==2.32.3
google-api-python-client==2.154.0
google-auth-oauthlib==1.2.1
playwright==1.49.0
aiofiles==24.1.0
python-multipart==0.0.19
jinja2==3.1.4
numpy>=1.26.0
pyttsx3==2.90
